Make Polux bushes check correct map for artificial building
Currently, they check the current map. This will work as long as the player is on a single map - in case of multiple settlements, or doing stuff off-map (attacking hostiles, quests, etc.) - this could cause the Polux bushes to to stop removing pollution when there's no artificial buildings, or start doing so even though there are some.